Rough code for first part of iteration 3
This is just an idea for part one, it said we should use the length() method (already in java) to figure out which movie_number entries need to be updated.  Works fine, I think we should remember that the length() method is probably going to include spaces (or other) so it won't match with the logical length using the API the instructor gave us.   Slow as heck, but it works.  
The insert part of this code works fine too but it wasn't easy trying to figure out how update SQL with just one ResultSet object open at a time.  That is why I put it all into an array.